requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
optuna>=3.4
torch>=2.1
matplotlib>=3.7
seaborn>=0.13
pyyaml>=6.0
rich>=13.7
tqdm>=4.66
